requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dataclasses-json==0.6.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.4
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.300
langsmith==0.0.40
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.6
numpy==1.26.0
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.1.1
pandas-stubs==2.0.3.230814
Pillow==10.0.1
plotly==5.17.0
portalocker==2.8.2
protobuf==4.24.3
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qdrant-client==1.5.4
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
types-pytz==2023.3.1.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
yarl==1.9.2
zipp==3.17.0
